+++ CHANGELOG.md
@@ -33,6 +33,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- [Oban.Telemetry] The format for job telemetry has changed to match the new
+  telemetry `span` convention.
+
+    * `[:oban, :started]` -> `[:oban, :job, :start]`
+    * `[:oban, :success]` -> `[:oban, :job, :stop]`
+    * `[:oban, :failure]` -> `[:oban, :job, :exception]`
+
+  In addition, for exceptions the stacktrace meta key has changed from `:stack`
+  to the standardized `:stactrace`.
+
 - [Oban.Notifier] Make the module public and clean up the primary function
   interfaces. Listening for and delivering notifications is simplified and no
   longer requires macros for pattern matching.